requirements.txt
streamlit>=1.31.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
openai>=1.12.0
python-dotenv>=1.0.0
phidata==2.3.5
duckduckgo-search==4.4.3
ta==0.11.0  # Technical Analysis library 